+++ src/ClickEvent.cpp
@@ -4,6 +4,28 @@
 
 #include "App.hpp"
 
+#include "Util/Logger.hpp"
+#include "eventpp/utilities/argumentadapter.h"
+#include "eventpp/utilities/conditionalfunctor.h"
+
+#include <Util/Input.hpp>
+#include <algorithm>
+#include "Actions.h"
+#include "Display/BeatHeart.h"
+#include "Display/BeatIndicator.h"
+#include "Dungeon/Enemy.h"
+#include "Dungeon/MapHandler.h"
+#include "Event/Event.h"
+#include "Game/Graphs/Spear.h"
+#include "Game_config.h"
+#include "HPIS.h"
+#include "Helper.hpp"
+#include "Music/Player.h"
+#include "Music/Tempo.h"
+#include "System.h"
+#include "Systems/HEIS.h"
+#include "Systems/HPIS.h"
+
 struct ClickEventType {
     std::vector<Util::Keycode> code;
     std::function<void()>      fptr;
@@ -11,13 +33,191 @@ struct ClickEventType {
 
 using CET = ClickEventType;
 
-auto printT = ClickEventType{{Util::Keycode::T}, []() { LOG_INFO("T"); }};
-
-auto test2 = CET{{Util::Keycode::RIGHT, Util::Keycode::LEFT}, []() {
-                     LOG_INFO("cond");
-                 }};
+auto musicTime = []() {
+    return static_cast<std::size_t>(Music::Player::GetMusicTime() * 1000)
+           % static_cast<std::size_t>(Music::Player::GetMusicLength() * 1000);
+};
 
 void App::BeforeUpdate() {
-    m_EventHandler.AddEvent(printT.code, printT.fptr);
-    m_EventHandler.AddEvent(test2.code, test2.fptr);
+    m_EventHandler.AddEvent(
+        [this]() {
+            LOG_INFO("aa");
+
+            if (!m_ThrowMode
+                || !Music::Tempo::IsTempoInRange(
+                    500,
+                    musicTime(),
+                    Music::Player::LoopCounter()
+                )) {
+                return;
+            }
+
+            std::for_each(
+                m_MapTableCodeDire.begin(),
+                m_MapTableCodeDire.end(),
+                [](const auto& elem) {
+                    if (!Util::Input::IsKeyDown(elem.first)) {
+                        return;
+                    }
+
+                    const auto& imagePath = m_MainCharacter->GetToolMod()
+                                                ->GetTool<IEquip>(3, "Spear")
+                                                ->GetImagePath();
+
+                    Game::Actions::ThrowOutWeapon(
+                        m_DungeonMap.get(),
+                        elem.second,
+                        imagePath
+                    );
+                }
+            );
+
+            m_ThrowMode = false;
+            m_MainCharacter
+                ->MoveByTime(200, m_AniPlayerDestination, m_PlayerMoveDirect);
+            m_MainCharacter->Update();
+            m_Camera->MoveByTime(200, m_AniCameraDestination);
+            m_DungeonMap->PlayerTrigger();
+        },
+        Util::Keycode::W,
+        Util::Keycode::D,
+        Util::Keycode::S,
+        Util::Keycode::A
+    );
+
+    m_EventHandler.AddEvent(
+        []() {
+            LOG_INFO("Throw Mode");
+            m_MainCharacter->PrepareThrowOut(true);
+            m_ThrowMode = true;
+        },
+        {Util::Keycode::UP, Util::Keycode::DOWN}
+    );
+
+    m_EventHandler.AddEvent(
+        []() {
+            m_DungeonMap->LoadLevel(m_DungeonMap->GetLevelNum() + 1);
+            m_AniCameraDestination = {0, 0};
+            m_AniPlayerDestination = {0, 0};
+        },
+        Util::Keycode::N
+    );
+
+    m_EventHandler.AddEvent(
+        [this]() { m_CurrentState = State::END; },
+        Util::Keycode::ESCAPE
+    );
+
+    m_EventHandler.AddEvent(
+        [this]() {
+            LOG_INFO("test");
+
+            if (m_ThrowMode
+                || !Music::Tempo::IsTempoInRange(
+                    500,
+                    musicTime(),
+                    Music::Player::LoopCounter()
+                )) {
+                return;
+            }
+
+            glm::vec2 playerDestination = m_MainCharacter->GetGamePosition();
+
+            if (m_PlayerMoveDirect != Player::NONE) {
+                m_PlayerMoveDirect = Player::NONE;
+            }
+            const std::vector<Util::Keycode> key = {
+              Util::Keycode::W,
+              Util::Keycode::A,
+              Util::Keycode::S,
+              Util::Keycode::D};
+            const std::vector<glm::vec2> direction =
+                {{0, -1}, {-1, 0}, {0, 1}, {1, 0}};
+            const std::vector<Player::Direction> playerDirection = {
+              Player::Direction::UP,
+              Player::Direction::LEFT,
+              Player::Direction::DOWN,
+              Player::Direction::RIGHT};
+            const std::vector<glm::vec2> aniPlayerDirection = {
+              {0, DUNGEON_TILE_WIDTH * DUNGEON_SCALE},
+              {-DUNGEON_TILE_WIDTH * DUNGEON_SCALE, 0},
+              {0, -DUNGEON_TILE_WIDTH * DUNGEON_SCALE},
+              {DUNGEON_TILE_WIDTH * DUNGEON_SCALE, 0}};
+            const std::vector<glm::vec2> aniCameraDirection = {
+              {0, -DUNGEON_TILE_WIDTH * DUNGEON_SCALE},
+              {DUNGEON_TILE_WIDTH * DUNGEON_SCALE, 0},
+              {0, DUNGEON_TILE_WIDTH * DUNGEON_SCALE},
+              {-DUNGEON_TILE_WIDTH * DUNGEON_SCALE, 0}};
+
+            for (std::size_t i = 0; i < 4; i++) {
+                if (Util::Input::IsKeyDown(key[i])
+                    && m_DungeonMap->GetMapData()->IsPositionPlayerAct(
+                        m_MainCharacter->GetGamePosition() + direction[i]
+                    )) {
+                    // origin mapdata actions
+                    playerDestination = m_MainCharacter->GetGamePosition()
+                                        + direction[i];
+                    m_MainCharacter->SetFaceTo(playerDirection[i]);
+                    auto mapIndex = m_DungeonMap->GamePostion2MapIndex(
+                        playerDestination
+                    );
+                    if (m_DungeonMap->GetMapData()->IsPositionInteractive(
+                            playerDestination
+                        )) {
+                        if (m_DungeonMap->GetMapData()->IsEnemyEmpty(mapIndex
+                            )) {
+                            m_DungeonMap->GetMapData()
+                                ->GetEnemy(mapIndex)
+                                ->Struck(2);
+
+                            Game::Systems::HEIS::DetectHealth(mapIndex);
+
+                            m_Camera->Shake(150, 10);
+                        } else if (m_DungeonMap->GetMapData()->IsPositionWall(
+                                       playerDestination
+                                   )) {
+                            m_DungeonMap->RemoveWall(
+                                m_DungeonMap->GamePostion2MapIndex(
+                                    playerDestination
+                                )
+                            );
+                        } else if (m_DungeonMap->GetMapData()->IsPositionDoor(
+                                       playerDestination
+                                   )) {
+                            m_DungeonMap->OpenDoor(
+                                m_DungeonMap->GamePostion2MapIndex(
+                                    playerDestination
+                                )
+                            );
+                            m_Camera->Shake(100, 10);
+                        }
+                    } else {
+                        m_PlayerMoveDirect = playerDirection[i];
+
+                        m_AniPlayerDestination = {
+                          m_AniPlayerDestination.x + aniPlayerDirection[i].x,
+                          m_AniPlayerDestination.y + aniPlayerDirection[i].y};
+                        m_AniCameraDestination = {
+                          m_AniCameraDestination.x + aniCameraDirection[i].x,
+                          m_AniCameraDestination.y + aniCameraDirection[i].y};
+                    }
+                }
+            }
+            m_MainCharacter
+                ->MoveByTime(200, m_AniPlayerDestination, m_PlayerMoveDirect);
+            m_MainCharacter->Update();
+            Event::EventQueue.dispatch(
+                m_MainCharacter.get(),
+                EventArgs(EventType::PlayerMove)
+            );
+            m_Camera->MoveByTime(200, m_AniCameraDestination);
+            m_DungeonMap->PlayerTrigger();
+        },
+        Util::Keycode::W,
+        Util::Keycode::D,
+        Util::Keycode::S,
+        Util::Keycode::A
+    );
+
+    //    m_EventHandler.AddEvent({Util::Keycode::W}, ThrowWeapon);
 };
