+++ src/main.cpp
@@ -10,12 +10,12 @@ int main(int, char**) {
 
     while (!context->GetExit()) {
         switch (app.GetCurrentState()) {
-        case App::State::START: app.Start(); break;
-
-        case App::State::UPDATE:
-            app.BeforeUpdate();
-            app.Update();
+        case App::State::START:
+            app.Start();
+            //            app.BeforeUpdate();
             break;
+
+        case App::State::UPDATE: app.Update(); break;
 
         case App::State::END:
             app.End();
